main.py
"""
[PROJECT NAME]
[One-line description]

Usage:
    python main.py --input <input_path> --output <output_path>
"""

import argparse
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()


def main(input_path: str, output_path: str) -> None:
    """
    Main entry point.
    
    Args:
        input_path: Path to input file or directory
        output_path: Path to output directory
    """
    # TODO: Implement core logic
    print(f"Processing: {input_path}")
    print(f"Output to: {output_path}")
    
    # 1. Load/process input
    # 2. Run AI processing
    # 3. Verify outputs
    # 4. Apply supervision rules
    # 5. Save results
    
    pass


if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="[PROJECT DESCRIPTION]")
    parser.add_argument("--input", "-i", required=True, help="Input file or directory")
    parser.add_argument("--output", "-o", default="results/", help="Output directory")
    
    args = parser.parse_args()
    
    # Create output directory if needed
    os.makedirs(args.output, exist_ok=True)
    
    main(args.input, args.output)
